Delegate responsibility for workflows to new class
