+++ ironflow/gui/gui.py
@@ -11,21 +11,13 @@ from typing import TYPE_CHECKING, Optional, Type
 
 import ipywidgets as widgets
 
-from ironflow.gui.workflows.boxes.flow import FlowBox
-from ironflow.gui.workflows.boxes.node_interface.control import NodeController
-from ironflow.gui.workflows.boxes.node_interface.representation import NodePresenter
-from ironflow.gui.workflows.boxes.text_output import TextOut
-from ironflow.gui.workflows.boxes.toolbar import Toolbar
-from ironflow.gui.workflows.boxes.user_input import UserInput
 from ironflow.gui.browser import Browser
-from ironflow.gui.workflows.canvas_widgets.flow import FlowCanvas
 from ironflow.gui.log import LogScreen
+from ironflow.gui.workflows.screen import WorkflowsScreen
 from ironflow.model.model import HasSession
-from ironflow.utils import display_string
 
 if TYPE_CHECKING:
     from ironflow.model.node import Node
-    from ironflow.gui.workflows.canvas_widgets.nodes import NodeWidget
 
 
 class GUI(HasSession):
@@ -74,13 +66,7 @@ class GUI(HasSession):
             enable_ryven_log=enable_ryven_log,
         )
 
-        self.flow_canvases = []
-        self.toolbar = Toolbar()
-        self.node_controller = NodeController(self)
-        self.node_presenter = NodePresenter()
-        self.text_out = TextOut()
-        self.input = UserInput()
-        self.flow_box = FlowBox(self.nodes_dictionary)
+        self.workflows_screen = WorkflowsScreen(self)
         self.browser = Browser()
 
         try:
@@ -91,7 +77,7 @@ class GUI(HasSession):
                 f"No session data found for {self.session_title}, making a new script."
             )
             self.create_script(script_title)
-        self.update_tabs()
+        self.workflows_screen.update_tabs()
 
     def create_script(
         self,
@@ -102,26 +88,15 @@ class GUI(HasSession):
         super().create_script(
             title=title, create_default_logs=create_default_logs, data=data
         )
-        self.flow_canvases.append(FlowCanvas(gui=self))
-
-    def delete_script(self) -> None:
-        self.flow_canvases.pop(self.active_script_index)
-        self.node_controller.close()
-        self.node_presenter.close()
-        super().delete_script()
-
-    @property
-    def flow_canvas(self):
-        return self.flow_canvases[self.active_script_index]
+        self.workflows_screen.add_flow(self.flow)
 
     @property
     def selected_node(self) -> Node | None:
-        selected = self.flow_canvas.get_selected_objects()
-        return selected[0].node if len(selected) > 0 else None
+        return self.workflows_screen.selected_node
 
     @property
     def new_node_class(self):
-        return self.flow_box.node_selector.new_node_class
+        return self.workflows_screen.new_node_class
 
     def serialize(self) -> dict:
         data = super().serialize()
@@ -129,7 +104,7 @@ class GUI(HasSession):
         for i_script, script in enumerate(self.session.scripts):
             all_data = data["scripts"][i_script]["flow"]["nodes"]
             self.active_script_index = i_script
-            for i, node_widget in enumerate(self.flow_canvas.objects_to_draw):
+            for i, node_widget in enumerate(self.workflows_screen.flow_canvas.objects_to_draw):
                 all_data[i]["pos x"] = node_widget.x
                 all_data[i]["pos y"] = node_widget.y
         self.active_script_index = currently_active
@@ -137,67 +112,15 @@ class GUI(HasSession):
 
     def load_from_data(self, data: dict) -> None:
         super().load_from_data(data)
-        self.flow_canvases = []
-        for i_script, script in enumerate(self.session.scripts):
-            flow_canvas = FlowCanvas(gui=self, flow=script.flow)
-            all_data = data["scripts"][i_script]["flow"]["nodes"]
-            for i_node, node in enumerate(script.flow.nodes):
-                flow_canvas.load_node(
-                    all_data[i_node]["pos x"], all_data[i_node]["pos y"], node
-                )
-            flow_canvas._built_object_to_gui_dict()
-            flow_canvas.redraw()
-            self.flow_canvases.append(flow_canvas)
+        self.workflows_screen.load_from_data(data)
 
     def register_node(self, node_class: Type[Node], node_group: Optional[str] = None):
         # Inherited __doc__ still applies just fine, all we do here is update a menu item afterwards.
         super().register_node(node_class=node_class, node_group=node_group)
         try:
-            self.flow_box.node_selector.update(self.nodes_dictionary)
+            self.workflows_screen.update_nodes_selector(self.nodes_dictionary)
         except AttributeError:
             pass  # It's not defined yet in the super().__init__ call, which is fine
-
-    def update_tabs(self):
-        self.flow_box.update_tabs(
-            outputs=[fc.output for fc in self.flow_canvases],
-            titles=[fc.title for fc in self.flow_canvases],
-            active_index=self.active_script_index,
-        )
-        for fc in self.flow_canvases:
-            fc.display()
-
-    def open_node_control(self, node: Node) -> None:
-        self.node_controller.draw_for_node(node)
-
-    def update_node_control(self) -> None:
-        self.node_controller.draw()
-
-    def close_node_control(self) -> None:
-        self.node_controller.node = None
-        self.node_controller.clear_output()
-
-    def ensure_node_not_controlled(self, node: Node) -> None:
-        if self.node_controller.node == node:
-            self.node_controller.draw_for_node(None)
-
-    def open_node_presenter(self, node_widget: NodeWidget):
-        self.node_presenter.node_widget = node_widget
-
-    def update_node_presenter(self):
-        self.node_presenter.draw()
-
-    def close_node_presenter(self):
-        self.node_presenter.close()
-
-    def ensure_node_not_presented(self, node_widget: NodeWidget) -> None:
-        if self.node_presenter.node_widget == node_widget:
-            self.node_presenter.node_widget = None
-
-    def redraw_active_flow_canvas(self):
-        self.flow_canvas.redraw()
-
-    def print(self, msg: str):
-        self.text_out.print(msg)
 
     def log_to_display(self):
         self.log_screen.log_to_display()
@@ -205,42 +128,15 @@ class GUI(HasSession):
     def log_to_stdout(self):
         self.log_screen.log_to_stdout()
 
-    def draw(self) -> widgets.VBox:
+    def draw(self) -> widgets.Tab:
         """
         Build the gui.
 
         Returns:
-            ipywidgets.VBox: The gui.
+            ipywidgets.Tab: The gui.
         """
 
-        # Wire callbacks
-        self.toolbar.alg_mode_dropdown.observe(
-            self._change_alg_mode_dropdown, names="value"
-        )
-        self.toolbar.buttons.help_node.on_click(self._click_node_help)
-        self.toolbar.buttons.load.on_click(self._click_load)
-        self.toolbar.buttons.save.on_click(self._click_save)
-        self.toolbar.buttons.add_node.on_click(self._click_add_node)
-        self.toolbar.buttons.delete_node.on_click(self._click_delete_node)
-        self.toolbar.buttons.create_script.on_click(self._click_create_script)
-        self.toolbar.buttons.rename_script.on_click(self._click_rename_script)
-        self.toolbar.buttons.delete_script.on_click(self._click_delete_script)
-        self.toolbar.buttons.zero_location.on_click(self._click_zero_location)
-        self.toolbar.buttons.zoom_in.on_click(self._click_zoom_in)
-        self.toolbar.buttons.zoom_out.on_click(self._click_zoom_out)
-        self.flow_box.script_tabs.observe(self._change_script_tabs)
-
-        flow_screen = widgets.VBox(
-            [
-                self.toolbar.box,
-                self.input.box,
-                self.flow_box.box,
-                self.text_out.box,
-                widgets.HBox([self.node_controller.box, self.node_presenter.box]),
-            ]
-        )
-
-        window = widgets.Tab([flow_screen, self.browser.box, self.log_screen.box])
+        window = widgets.Tab([self.workflows_screen.box, self.browser.box, self.log_screen.box])
         window.set_title(0, "Workflow")
         window.set_title(1, "Browser")
         window.set_title(2, "Log")
@@ -249,118 +145,6 @@ class GUI(HasSession):
 
     # Type hinting for unused `change` argument in callbacks taken from ipywidgets docs:
     # https://ipywidgets.readthedocs.io/en/latest/examples/Widget%20Events.html#Traitlet-events
-    def _change_alg_mode_dropdown(self, change: dict) -> None:
-        # Current behaviour: Updates the flow mode for all scripts
-        # Todo: Change only for the active script, and update the dropdown on tab (script) switching
-        for script in self.session.scripts:
-            script.flow.set_algorithm_mode(self.toolbar.alg_mode_dropdown.value)
-
-    def _click_save(self, change: dict) -> None:
-        self.input.open_text(
-            "Save file",
-            self._click_confirm_save,
-            self.session_title,
-            description_tooltip="Save to file name (omit the file extension, .json)",
-        )
-        self.print("Choose a file name to save to (omit the file extension, .json)")
-
-    def _click_confirm_save(self, change: dict) -> None:
-        file_name = self.input.text
-        self.save(f"{file_name}.json")
-        self.print(f"Session saved to {file_name}.json")
-        self.input.clear()
-
-    def _click_load(self, change: dict) -> None:
-        self.input.open_text(
-            "Load file",
-            self._click_confirm_load,
-            self.session_title,
-            description_tooltip="Load from file name (omit the file extension, .json).",
-        )
-        self.print("Choose a file name to load (omit the file extension, .json)")
-
-    def _click_confirm_load(self, change: dict) -> None:
-        file_name = self.input.text
-        self.load(f"{file_name}.json")
-        self.update_tabs()
-        self.node_presenter.clear_output()
-        self.print(f"Session loaded from {file_name}.json")
-        self.input.clear()
-
-    def _click_node_help(self, change: dict) -> None:
-        self.print(
-            display_string(
-                f"{self.new_node_class.__name__.replace('_Node', '')}:\n{self.new_node_class.__doc__}"
-            )
-        )
-
-    def _click_add_node(self, change: dict) -> None:
-        self.flow_canvas.add_node(10, 10, self.new_node_class)
-
-    def _click_delete_node(self, change: dict) -> None:
-        self.flow_canvas.delete_selected()
-
-    def _click_create_script(self, change: dict) -> None:
-        self.create_script()
-        self.update_tabs()
-
-    def _click_rename_script(self, change: dict) -> None:
-        self.input.open_text(
-            "New name",
-            self._click_confirm_rename,
-            self.script.title,
-            description_tooltip="New script name",
-        )
-        self.print("Choose a new name for the current script")
-
-    def _click_confirm_rename(self, change: dict) -> None:
-        new_name = self.input.text
-        old_name = self.script.title
-        rename_success = self.rename_script(new_name)
-        if rename_success:
-            self.flow_box.script_tabs.set_title(self.active_script_index, new_name)
-            self.print(f"Script '{old_name}' renamed '{new_name}'")
-        else:
-            self.print(
-                f"INVALID NAME: Failed to rename script '{self.script.title}' to '{new_name}'."
-            )
-
-    def _click_delete_script(self, change: dict) -> None:
-        self.input.open_bool(
-            f"Delete the entire script {self.script.title}?",
-            self._click_confirm_delete_script,
-        )
-
-    def _click_confirm_delete_script(self, change: dict) -> None:
-        script_name = self.script.title
-        self.delete_script()
-        self.update_tabs()
-        self.print(f"Script {script_name} deleted")
-
-    def _click_zero_location(self, change: dict) -> None:
-        self.flow_canvas.x = 0
-        self.flow_canvas.y = 0
-        self.flow_canvas.redraw()
-
-    def _click_zoom_in(self, change: dict) -> None:
-        self.flow_canvas.zoom_in()
-
-    def _click_zoom_out(self, change: dict) -> None:
-        self.flow_canvas.zoom_out()
-
-    def _click_input_text_cancel(self, change: dict) -> None:
-        self.input.clear()
-        self.text_out.clear()
-
-    def _change_script_tabs(self, change: dict):
-        if change["name"] == "selected_index" and change["new"] is not None:
-            self.input.clear()
-            self.flow_canvas.deselect_all()
-            if self.flow_box.script_tabs.selected_index == self.n_scripts:
-                self.create_script()
-                self.update_tabs()
-            else:
-                self.active_script_index = self.flow_box.script_tabs.selected_index
 
     def _change_screen_tabs(self, change: dict):
         if change["name"] == "selected_index" and change["new"] == 1:
