+++ ironflow/gui/workflows/canvas_widgets/base.py
@@ -16,8 +16,8 @@ from ironflow.gui.workflows.canvas_widgets.layouts import Layout
 
 Number = Union[int, float]
 if TYPE_CHECKING:
-    from ironflow.gui.gui import GUI
     from ironflow.gui.workflows.canvas_widgets.flow import FlowCanvas
+    from ironflow.gui.workflows.screen import WorkflowsScreen
     from ironflow.model import Flow
 
 
@@ -77,8 +77,8 @@ class CanvasWidget(ABC):
         return self.parent.canvas
 
     @property
-    def gui(self) -> GUI:
-        return self.parent.gui
+    def screen(self) -> WorkflowsScreen:
+        return self.parent.screen
 
     @property
     def flow(self) -> Flow:
