+++ tests/unit/test_gui.py
@@ -18,9 +18,9 @@ class TestGUI(TestCase):
 
     def test_multiple_scripts(self):
         gui = GUI('foo')
-        gui.flow_canvas.add_node(0, 0, gui.nodes_dictionary['built_in']['val'])
+        gui.workflows_screen.flow_canvas.add_node(0, 0, gui.nodes_dictionary['built_in']['val'])
         gui.create_script()
-        gui.flow_canvas.add_node(1, 1, gui.nodes_dictionary['built_in']['result'])
+        gui.workflows_screen.flow_canvas.add_node(1, 1, gui.nodes_dictionary['built_in']['result'])
         canonical_file_name = f"{gui.session_title}.json"
         gui.save(canonical_file_name)
         new_gui = GUI('something_random')
@@ -46,7 +46,7 @@ class TestGUI(TestCase):
     def test_saving_and_loading(self):
         title = 'foo'
         gui = GUI(title)
-        canvas = gui.flow_canvas
+        canvas = gui.workflows_screen.flow_canvas
         flow = gui._session.scripts[0].flow
 
         self.assertEqual(0, len(flow.nodes), msg="Fresh GUI shouldn't have any nodes yet.")
@@ -113,7 +113,7 @@ class TestGUI(TestCase):
         gui.register_node(MyNode, node_group="user")
         self.assertIn(MyNode, gui.session.nodes)
 
-        gui.flow_canvas.add_node(0, 0, gui.nodes_dictionary["user"][MyNode.title])
+        gui.workflows_screen.flow_canvas.add_node(0, 0, gui.nodes_dictionary["user"][MyNode.title])
         gui.flow.nodes[0].inputs[0].update(1)
         self.assertEqual(gui.flow.nodes[0].outputs[0].val, 43)
 
@@ -134,7 +134,7 @@ class TestGUI(TestCase):
         gui.flow.nodes[0].inputs[0].update(2)
         self.assertEqual(gui.flow.nodes[0].outputs[0].val, 44, msg="Expected to be using instance of old class")
 
-        gui.flow_canvas.add_node(1, 1, gui.nodes_dictionary["user"][MyNode.title])
+        gui.workflows_screen.flow_canvas.add_node(1, 1, gui.nodes_dictionary["user"][MyNode.title])
         gui.flow.nodes[1].inputs[0].update(2)
         self.assertEqual(gui.flow.nodes[1].outputs[0].val, -40, msg="New node instances should reflect updated class.")
